Implementacao da atualizacao de status de pagamento com o RabbitMQ

+++ src/GeekShopping.OrderAPI/Model/OrderHeader.cs
@@ -11,9 +11,9 @@ namespace GeekShopping.OrderAPI.Model
         [Column("coupon_code")]
         public string CouponCode { get; set; }
         [Column("purchase_amount")] 
-        public double PurchaseAmount { get; set; }
+        public decimal PurchaseAmount { get; set; }
         [Column("discount_amount")] 
-        public double DiscountAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
         [Column("first_name")] 
         public string FirstName { get; set; }
         [Column("last_name")]
